qc/qualimap: relocate to 'qc.apps.qualimap'.
