+++ auto_process_ngs/qc/reporting.py
@@ -114,6 +114,7 @@ from .apps.fastq_strand import fastq_strand_output
 from .apps.cellranger import CellrangerCount
 from .apps.cellranger import CellrangerMulti
 from .apps.picard import CollectInsertSizeMetrics
+from .apps.qualimap import QualimapRnaseq
 from .apps.seqlens import SeqLens
 from .outputs import QCOutputs
 from .outputs import ExtraOutputs
@@ -133,7 +134,6 @@ from .plots import ugenomicoriginplot
 from .plots import encode_png
 from .protocols import QCProtocol
 from .protocols import fetch_protocol_definition
-from .qualimap import QualimapRnaseq
 from .rseqc import InferExperiment
 from ..settings import get_install_dir
 from .utils import get_bam_basename
